requirement.txt
google-cloud-vision 
google-cloud-pubsub 
google-cloud-storage